Multiple changes... See Description
Implementation of IdentityRole for Users, and ability to assign users to a role when they log in.

Test code in HomeController that works for Uploading files to blob

Added PartnerMembers table, which will map users to their respective organization

Added another line to all addresses.... Address line 2. This will be used for Apartment numbers, building numbers, etc.

Created validation for file types. Primarily will be used for Partner Logo uploads. See //ViewModels/Home/Upload file for example of how it is used

Added further requirements to Requrement doc that I am currently working on.

Added signInManager to _ViewImports

Partners now have string logo that will be used to locate where the Partners' logo is located. This will primarily be used on Opportunity pages, and the Partner Page.

Will be adding an adminController that will mostly provide Admin functionality to Viewing Logs etc.